move DUR middleware to middleware folder
